split long testfile
